api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
alembic
python-dotenv
passlib
bcrypt==3.2.0
python-jose[cryptography]
psycopg2-binary
requests
loguru>=0.7.0
pydantic[email]
python-multipart
fpdf2
reportlab
fastapi-mail
twilio